+++ NEWS.md
@@ -1,6 +1,6 @@
 
 # OpenRepGrid 0.1.11 (not yet released) 
-
+  * saveAsExcel to save grids as Microsoft Excel files
   * replace xlsx by openxlsx to import Excel files to get rid of JRE dependency
   
 # OpenRepGrid 0.1.10 (released February 24, 2017)